logic classes organized
