+++ backend/api_tests/debug.txt
@@ -1,1 +1,1 @@
-act_put_crit
+eupn_signup
