+++ backend/logic_classes/helpers/google_auth_extract.py
