+++ backend/logic_classes/helpers/user_name_flatten.py
